feat(test): add hash verification for downloaded files
Introduce a utility function `file2sha256` in the test module to calculate
the SHA256 hash of a file. This enables the verification of the integrity of
downloaded files by comparing the calculated hash with an expected value.

The hash calculation is performed in chunks to optimize memory usage, reading
the file in1024-byte blocks and updating the hash digest accordingly.