fix(yundownload): adjust chunk sizes for improved download efficiency
Increase the STREAM_SIZE in the YunDownloader class to1MB to enhance the
asynchronous writing of files during downloads. This change ensures that
the downloads are processed more efficiently with larger chunks.

Additionally, update the version in pyproject.toml to 0.2.7 to reflect the
codebase changes. The README is also updated to document the bug fixes in
this version.